# bee_agents/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scholarship Analysis API",
    "description": "REST API for accessing scholarship application analysis data with multi-tenancy support",
    "version": "2.0.0"
  },
  "paths": {
...
    "/application": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Application Analysis",
        "description": "Get detailed application analysis.\n\nArgs:\n    scholarship: Name of the scholarship\n    wai_number: WAI application number\n    \nReturns:\n    ApplicationAnalysisResponse with detailed analysis",
        "operationId": "get_application",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')"
          },
          {
            "name": "wai_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "WAI application number",
              "title": "Wai Number"
            },
            "description": "WAI application number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/academic": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Academic Analysis",
        "description": "Get academic analysis for an application.\n\nArgs:\n    scholarship: Name of the scholarship\n    wai_number: WAI application number\n    \nReturns:\n    AcademicAnalysisResponse with academic analysis",
        "operationId": "get_academic",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')"
          },
          {
            "name": "wai_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "WAI application number",
              "title": "Wai Number"
            },
            "description": "WAI application number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/essay": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Single Essay Analysis",
        "description": "Get analysis for a specific essay.\n\nArgs:\n    essay_number: Essay number (1 or 2)\n    scholarship: Name of the scholarship\n    wai_number: WAI application number\n    \nReturns:\n    EssayAnalysisResponse for the specified essay",
        "operationId": "get_essay",
        "parameters": [
          {
            "name": "essay_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Essay number (1 or 2)",
              "title": "Essay Number"
            },
            "description": "Essay number (1 or 2)",
            "example": 1
          },
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')"
          },
          {
            "name": "wai_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "WAI application number",
              "title": "Wai Number"
            },
            "description": "WAI application number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EssayAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendation": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Recommendation Analysis",
        "description": "Get combined recommendation analysis for an application.\n\nArgs:\n    scholarship: Name of the scholarship\n    wai_number: WAI application number\n    \nReturns:\n    Combined analysis of all recommendations",
        "operationId": "list_recs",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')"
          },
          {
            "name": "wai_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "WAI application number",
              "title": "Wai Number"
            },
            "description": "WAI application number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendation/by-number": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Single Recommendation Analysis",
        "description": "Get analysis for a specific recommendation.\n\nArgs:\n    rec_number: Recommendation number (1 or 2)\n    scholarship: Name of the scholarship\n    wai_number: WAI application number\n    \nReturns:\n    RecommendationAnalysisResponse for the specified recommendation",
        "operationId": "get_recommendation",
        "parameters": [
          {
            "name": "rec_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Recommendation number (1 or 2)",
              "title": "Rec Number"
            },
            "description": "Recommendation number (1 or 2)",
            "example": 1
          },
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')"
          },
          {
            "name": "wai_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "WAI application number",
              "title": "Wai Number"
            },
            "description": "WAI application number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcademicAnalysisResponse": {
        "properties": {
          "wai_number": {
            "type": "string",
            "title": "Wai Number",
            "description": "WAI application number"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Summary of academic performance"
          },
          "profile_features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile Features",
            "description": "Academic profile features"
          },
          "scores": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scores",
            "description": "Academic scores"
          },
          "score_breakdown": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Breakdown",
            "description": "Score reasoning"
          },
          "processed_date": {
            "type": "string",
            "title": "Processed Date",
            "description": "When the analysis was processed"
          },
          "source_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File",
            "description": "Source file"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used",
            "description": "Model used for analysis"
          }
        },
        "type": "object",
        "required": [
          "wai_number",
          "summary",
          "profile_features",
          "processed_date"
        ],
        "title": "AcademicAnalysisResponse",
        "description": "Response model for academic analysis."
      },
      "ApplicationAnalysisResponse": {
        "properties": {
          "wai_number": {
            "type": "string",
            "title": "Wai Number",
            "description": "WAI application number"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Summary of application quality"
          },
          "scores": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Scores",
            "description": "Score breakdown"
          },
          "score_breakdown": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Score Breakdown",
            "description": "Detailed reasoning for scores"
          },
          "completeness_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Completeness Issues",
            "description": "List of completeness issues"
          },
          "validity_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Validity Issues",
            "description": "List of validity issues"
          },
          "attachment_status": {
            "type": "string",
            "title": "Attachment Status",
            "description": "Summary of attachment completeness"
          },
          "processed_date": {
            "type": "string",
            "title": "Processed Date",
            "description": "When the analysis was processed"
          },
          "source_file": {
            "type": "string",
            "title": "Source File",
            "description": "Source application file"
          }
        },
        "type": "object",
        "required": [
          "wai_number",
          "summary",
          "scores",
          "score_breakdown",
          "completeness_issues",
          "validity_issues",
          "attachment_status",
          "processed_date",
          "source_file"
        ],
        "title": "ApplicationAnalysisResponse",
        "description": "Response model for application analysis."
      },
...
      "EssayAnalysisResponse": {
        "properties": {
          "wai_number": {
            "type": "string",
            "title": "Wai Number",
            "description": "WAI application number"
          },
          "essay_number": {
            "type": "integer",
            "title": "Essay Number",
            "description": "Essay number (1 or 2)"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Summary of essay quality"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths",
            "description": "Essay strengths"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weaknesses",
            "description": "Essay weaknesses"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "Essay score if available"
          },
          "processed_date": {
            "type": "string",
            "title": "Processed Date",
            "description": "When the analysis was processed"
          }
        },
        "type": "object",
        "required": [
          "wai_number",
          "essay_number",
          "summary",
          "strengths",
          "weaknesses",
          "processed_date"
        ],
        "title": "EssayAnalysisResponse",
        "description": "Response model for essay analysis."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RecommendationAnalysisResponse": {
        "properties": {
          "wai_number": {
            "type": "string",
            "title": "Wai Number",
            "description": "WAI application number"
          },
          "recommendation_number": {
            "type": "integer",
            "title": "Recommendation Number",
            "description": "Recommendation number (1 or 2)"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Summary of recommendation quality"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths",
            "description": "Recommendation strengths"
          },
          "concerns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Concerns",
            "description": "Recommendation concerns"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "Recommendation score if available"
          },
          "processed_date": {
            "type": "string",
            "title": "Processed Date",
            "description": "When the analysis was processed"
          }
        },
        "type": "object",
        "required": [
          "wai_number",
          "recommendation_number",
          "summary",
          "strengths",
          "concerns",
          "processed_date"
        ],
        "title": "RecommendationAnalysisResponse",
        "description": "Response model for recommendation analysis."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use 'Bearer &lt;token&gt;' from the chat /login endpoint."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8200",
      "description": "Development server (default port)"
    }
  ]
}